+++ packages/cli/src/controller/publish-controller.ts
@@ -8,6 +8,7 @@ import {parseAlgorandProjectManifest} from '@subql/common-algorand';
 import {parseSubstrateProjectManifest as parseAvalancheProjectManifest} from '@subql/common-avalanche';
 import {parseCosmosProjectManifest} from '@subql/common-cosmos';
 import {parseEthereumProjectManifest} from '@subql/common-ethereum';
+import {parseEthereumProjectManifest as parseFlareProjectManifest} from '@subql/common-flare';
 import {parseSubstrateProjectManifest, manifestIsV0_0_1} from '@subql/common-substrate';
 import {parseTerraProjectManifest} from '@subql/common-terra';
 import {FileReference} from '@subql/types';
@@ -56,7 +57,11 @@ export async function uploadToIpfs(projectPath: string, authToken: string, ipfsE
             try {
               manifest = parseEthereumProjectManifest(schema).asImpl;
             } catch (e) {
-              throw new Error('Unable to pass project manifest');
+              try {
+                manifest = parseFlareProjectManifest(schema).asImpl;
+              } catch (e) {
+                throw new Error('Unable to pass project manifest');
+              }
             }
           }
         }