+++ packages/node-core/src/utils/sync-helper.ts
@@ -1,8 +1,8 @@
 // Copyright 2020-2022 OnFinality Limited authors & contributors
 // SPDX-License-Identifier: Apache-2.0
 
-import { blake2AsHex } from '@polkadot/util-crypto';
-import {Utils} from 'sequelize';
+import {blake2AsHex} from '@polkadot/util-crypto';
+import {QueryTypes, Sequelize, Utils} from 'sequelize';
 
 export interface SmartTags {
   foreignKey?: string;
@@ -122,17 +122,23 @@ END;
 $$ LANGUAGE plpgsql;`;
 
 export function dropNotifyTrigger(schema: string, table: string): string {
-  const triggerName = makeTriggerName(schema, table);
+  const triggerName = makeTriggerName(schema, table, 'notify');
   return `DROP TRIGGER IF EXISTS "${triggerName}"
     ON "${schema}"."${table}";`;
 }
 
-export function getNotifyTriggers(): string {
-  return `select trigger_name as "triggerName", event_manipulation as "eventManipulation" from information_schema.triggers
-          WHERE trigger_name = :triggerName`;
+export async function getTriggers(sequelize: Sequelize, triggerName: string): Promise<any[]> {
+  return sequelize.query(
+    `select trigger_name as "triggerName", event_manipulation as "eventManipulation" from information_schema.triggers
+          WHERE trigger_name = :triggerName`,
+    {
+      replacements: {triggerName},
+      type: QueryTypes.SELECT,
+    }
+  );
 }
 export function createNotifyTrigger(schema: string, table: string): string {
-  const triggerName = makeTriggerName(schema, table);
+  const triggerName = makeTriggerName(schema, table, 'notify');
   return `
 CREATE TRIGGER "${triggerName}"
     AFTER INSERT OR UPDATE OR DELETE
@@ -140,7 +146,31 @@ CREATE TRIGGER "${triggerName}"
     FOR EACH ROW EXECUTE FUNCTION send_notification();`;
 }
 
-export function makeTriggerName(schema: string, tableName: string): string {
+export function makeTriggerName(schema: string, tableName: string, triggerType: string): string {
   // max name length is 63 bytes in Postgres
-  return blake2AsHex(`${schema}_${tableName}_notify_trigger`).substr(2, 10);
+  return blake2AsHex(`${schema}_${tableName}_${triggerType}_trigger`).substr(2, 10);
+}
+
+export function createSchemaTrigger(schema: string): string {
+  const triggerName = makeTriggerName(schema, '_metadata', 'schema');
+  return `
+  CREATE TRIGGER "${triggerName}"
+    AFTER UPDATE
+    ON "${schema}"."_metadata"
+    FOR EACH ROW
+    WHEN ( new.key = 'schemaMigrationCount')
+    EXECUTE FUNCTION "${schema}".schema_notification();`;
+}
+
+export function createSchemaTriggerFunction(schema: string): string {
+  return `
+  CREATE OR REPLACE FUNCTION "${schema}".schema_notification()
+    RETURNS trigger AS $$
+  BEGIN
+    PERFORM pg_notify(
+            CONCAT(TG_TABLE_SCHEMA,'.',TG_TABLE_NAME,'.','hot_schema'),
+            'schema_updated');
+    RETURN NULL;
+  END;
+  $$ LANGUAGE plpgsql;`;
 }