+++ packages/node/src/indexer/project.service.ts
@@ -192,6 +192,7 @@ export class ProjectService {
       'chainId',
       'processedBlockCount',
       'lastFinalizedVerifiedHeight',
+      'schemaMigrationCount',
     ] as const;
 
     const entries = await metadataRepo.findAll({
@@ -257,6 +258,10 @@ export class ProjectService {
         value: packageVersion,
       });
     }
+    if (!keyValue.schemaMigrationCount) {
+      await metadataRepo.upsert({ key: 'schemaMigrationCount', value: 0 });
+    }
+
     return metadataRepo;
   }
 
@@ -302,7 +307,6 @@ export class ProjectService {
     } else {
       startHeight = this.getStartBlockFromDataSources();
     }
-
     return startHeight;
   }