+++ packages/node-core/CHANGELOG.md
@@ -10,6 +10,15 @@ All logs must start with the format: [x.y.z] - yyyy-mm-dd
 
 ## [Unreleased]
 
+## [1.4.1] - 2022-11-15
+### Fixed
+- Hot fix for hot schema reload (#1404)
+
+## [1.4.0] - 2022-11-15
+### Added
+- Support for hot schema reload. (#1401)
+- Support for distinct query plugin. (#1274)
+
 ## [1.3.3] - 2022-11-09
 ### Added
 - Added retry method for handle fetch errors (#1386)