+++ packages/node/CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.14.1] - 2022-11-16
+- Patch release with @subql/node-core fix (#1404)
+
+## [1.14.0] - 2022-11-15
+### Added
+- Support hot schema reload (#1401)
+
 ## [1.13.3] - 2022-11-09
 ### Fixed
 - Fix issue with reindex missing bind (#1391)